QRStuff: rest of the new login stuff
Flask-Login has been implemented, each user has a unique `username`, which is the primary key for the `Users` table.
Each user also has their own password, encrypted with `bcrypt`. Flask-Bcrypt has been used for the hashing.
Email address is also an attribute, with unique constraint. An API key is also generated, hashed, and stored.
It is displayed to the user only once, on account creation. Full name is stored just for each identification, incase
people choose usernames which aren't easily recognizable.

All views which require a user to be logged in are marked with a `@login_required` decorator
For views which have a frontend, it will simply redirect you to the `/login` view and the `next` argument will help to
get you redirected back to the desired view.

In the case of an API call, there are 2 methods of authentication. You can either pass an Authorization header with the
API key, or you can pass a Credentials header containing a base64 encoded version of username|password.

For registration, there is `/register` endpoint, which can be accessed via a web browser, or can be directly POSTed to.
It expects 4 fields, namely `name`, `username`, `password`, `email`. Incase of an IntegrityError because of a unique
constraint violation or similar, it will return error 400 and a corresponding message.
In the case of success, it will return 200 and the API key for the user, which should be saved, as it will not be
displayed anywhere else again.

Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>