Hades: update registration function
Move `db.session.add()` outside of the try-except block - that's not
going to can an IntegrityError.

Also add 2 more rows, to the `access` table if newly registered member
is a part of TSG, i.e. give them access to `tsg` and `test_users` table
by default

Migration script to achieve this for existing DB by Desai:
https://del.dog/hades_access.sql

Signed-off-by: Akhil Narang <me@akhilnarang.dev>